% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCzinb_internal.R
\name{nb.regression.parseModel}
\alias{nb.regression.parseModel}
\title{Parse ZINB regression model}
\usage{
nb.regression.parseModel(alpha, A.mu)
}
\arguments{
\item{alpha}{the vectors of parameters c(a.mu) concatenated}

\item{A.mu}{matrix of the model (default=empty)}
}
\value{
A list with slot \code{logMu},
}
\description{
Given the parameters of a NB regression model, this function parses the
model and computes the vector of log(mu), and the dimensions of
the different components of the vector of parameters. See
\code{\code{nb.loglik.regression}} for details of the NB regression model
and its parameters.
}
\seealso{
\code{\code{nb.loglik.regression}}
}
\keyword{internal}
