% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCzinb_internal.R
\name{nb.wald}
\alias{nb.wald}
\title{Structure learning with negative binomial model using glm}
\usage{
nb.wald(X, maxcard, alpha, extend, nCores = 1)
}
\arguments{
\item{X}{the matrix of counts (n times p).}

\item{maxcard}{the uper bound of the cardinality of the conditional sets K}

\item{alpha}{the significant level of the tests}

\item{extend}{if TRUE it considers the union of the tests, otherwise it
considers the intersection.}
}
\value{
the estimated adjacency matrix of the graph.
}
\description{
This function estimates the adjacency matrix of a NB model given a matrix of
counts, using the glm.nb function.
}
