% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathg.R
\name{pathg}
\alias{pathg}
\title{Select and Extract Highly Expressed Genes in a Seurat Object}
\usage{
pathg(seurat_object, cell_type, top_n = NULL)
}
\arguments{
\item{seurat_object}{A Seurat object containing RNA expression data.}

\item{cell_type}{A character string specifying the cell type of interest.}

\item{top_n}{An integer indicating the number of top expressed genes to select.}
}
\value{
A character vector of the selected genes.
}
\description{
This function identifies and extracts the top \code{n} most highly expressed genes
for a given cell type in a Seurat object.
}
\examples{
\dontrun{
selected_genes <- pathg(my_seurat_object, cell_type = "T_cells", top_n = 20)
}
}
