% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCzinb_internal.R
\name{prediction_scores}
\alias{prediction_scores}
\title{Prediction scores for estimated graph}
\usage{
prediction_scores(trueG, estimatedG, type)
}
\arguments{
\item{trueG}{the adjacency matrix of the true graph.}

\item{estimatedG}{the adjacency matrix of the estimated graph.}

\item{type}{the type of the true graph, could be "UG" undirected graph,
or "DAG" directed acyclic graph.}
}
\value{
A vector with the number of true positives (TP), the number of false
positives (FP), the number of false negatives (FN), the positive predictive
value (PPV), the sensitivity (Se), and the F1 score (F1).
}
\description{
This function returns a set of metrics useful to evaluate the goodness of the
estimation of the graph on simulated data, or whenever a true graph is
available.
}
\examples{
set.seed(123)
adj <- matrix(c(0, 1, 1, 1, 0, 1, 1, 1, 0), nrow = 3)
mat <- zinb_simdata(n = 100, p = 3, B = adj, 
                    mu_range = list(c(1, 5)), mu_noise = 1, 
                    theta = 1, pi = 0.1)[[1]]
res <- PCzinb(mat, method = "poi", alpha = 0.05)
prediction_scores(adj, res, type = "UG")
}
