% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_adj_matrix.R
\docType{data}
\name{toy_adj_matrix}
\alias{toy_adj_matrix}
\title{Toy adjacency matrix for examples}
\format{
A matrix of dimension 10x10.
}
\usage{
data(toy_adj_matrix)
}
\description{
An adjacency matrix generated using real dataset and STRINGdb for
demonstrating functions in the scGraphVerse package.
}
\examples{
data(toy_adj_matrix)
str(toy_adj_matrix)
}
\keyword{datasets}
