% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_truth.R
\docType{data}
\name{adj_truth}
\alias{adj_truth}
\title{Toy adjacency matrix for examples}
\format{
A matrix of dimension 35x35.
}
\usage{
data(adj_truth)
}
\description{
An adjacency matrix generated using real dataset and STRINGdb for
demonstrating functions in the scGraphVerse package.
}
\examples{
data(adj_truth)
str(adj_truth)
}
\keyword{datasets}
