% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_consensus.R
\name{compare_consensus}
\alias{compare_consensus}
\title{Compare Consensus and Original Graphs}
\usage{
compare_consensus(consensus_matrix, original_matrix)
}
\arguments{
\item{consensus_matrix}{A binary adjacency matrix representing the consensus graph.}

\item{original_matrix}{A binary adjacency matrix representing the original or ground truth graph.}
}
\value{
This function does not return any values. It generates two plots:
- One for true positives and false negatives.
- Another for false positives.
}
\description{
This function compares a consensus adjacency matrix with an original adjacency matrix, visualizing the true
positives (TP), false negatives (FN), and false positives (FP) between the two graphs. It uses the
\code{ggraph} package to create visualizations that show the agreement and disagreement between the two graphs.
The function generates two plots:
\enumerate{
\item A plot showing the edges present in both the ground truth and consensus matrix (TP) and the edges
present only in the ground truth (FN).
\item A plot showing the edges present in the consensus matrix but absent in the ground truth (FP).
}
}
\details{
The function compares two graphs: the consensus graph and the ground truth graph. It computes:
\itemize{
\item True Positives (TP): Edges that are present in both graphs.
\item False Negatives (FN): Edges that are present in the original graph but not in the consensus graph.
\item False Positives (FP): Edges that are present in the consensus graph but not in the original graph.
}

The true positives (TP) and false negatives (FN) are plotted together, while the false positives (FP) are plotted separately.
}
