% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_consensus.R
\name{compare_consensus}
\alias{compare_consensus}
\title{Compare Consensus and Original Graphs}
\usage{
compare_consensus(consensus_matrix, original_matrix)
}
\arguments{
\item{consensus_matrix}{A binary adjacency matrix representing the consensus graph.}

\item{original_matrix}{A binary adjacency matrix representing the original or ground truth graph.}
}
\value{
This function does not return any values. It generates two plots:
- The first for True Positives (red) and False Positives (blue).
- The second for False Negatives (purple).
}
\description{
This function compares a consensus adjacency matrix with an original adjacency matrix, visualizing edges as:
\itemize{
\item True Positives (TP): edges that are 1 in both (red).
\item False Positives (FP): edges that are 1 in the original but 0 in the consensus (blue).
\item False Negatives (FN): edges that are 0 in the original but 1 in the consensus (purple).
}
}
\details{
It generates two plots:
\enumerate{
\item A plot of the original (ground truth) graph with TP edges in red and FP edges in blue.
\item A plot showing only the FN edges (purple).
}
}
