% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrices.R
\docType{data}
\name{count_matrices}
\alias{count_matrices}
\title{Toy list of SingleCellExperiment objects}
\format{
A list with 3 SingleCellExperiment objects,
each of dimension 10 genes x 10 cells.
}
\usage{
data(count_matrices)
}
\description{
A simulated dataset generated using zinb_simdata() for demonstrating
functions in the scGraphVerse package.
}
\examples{
data(count_matrices)
str(count_matrices)
}
\keyword{datasets}
