% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consensus.R
\name{create_consensus}
\alias{create_consensus}
\title{Generate Consensus Adjacency Matrix from a List of Adjacency Matrices}
\usage{
create_consensus(
  adj_matrix_list,
  method = "vote",
  weighted_list = NULL,
  theta = 0.04,
  threshold = 0.5,
  ncores = 1
)
}
\arguments{
\item{adj_matrix_list}{A list of binary adjacency matrices to be used to generate the consensus matrix.
All matrices should have the same dimensions.}

\item{method}{A character string specifying the method for generating the consensus. Options are:
"vote" (default), "union", and "INet".}

\item{weighted_list}{A list of weighted adjacency matrices, required only when \code{method = "INet"}. It is
used to apply weighted information to the consensus generation.}

\item{theta}{A numeric value specifying the threshold for consensus when using the "INet" method. It is
passed to the \code{consensusNet} function. The default value is 0.04.}

\item{threshold}{A numeric value between 0 and 1. For the "vote" method, this threshold defines the minimum
percentage of matrices in which an edge must be present to be included in the consensus.
For the "union" method, any edge with a value >= 1 in any matrix will be included.}

\item{ncores}{The number of CPU cores to use when generating the consensus matrix using the "INet" method.
The default is 1 (single-core).}
}
\value{
A square matrix representing the consensus adjacency matrix.
}
\description{
This function generates a consensus adjacency matrix from a list of adjacency matrices. It supports
three methods for generating the consensus: "vote", "union", and "INet".
\itemize{
\item \strong{"vote"}: The consensus matrix is generated by selecting edges that are present in at least 75\% of the
matrices in the input list.
\item \strong{"union"}: The consensus matrix is generated by taking the union of all matrices, where any edge present
in any matrix is included in the consensus.
\item \strong{"INet"}: The consensus matrix is generated using weighted adjacency matrices, where the weights are
normalized and combined to form the consensus network.
}
}
\details{
The \code{create_consensus} function generates a consensus adjacency matrix from a list of binary or weighted
adjacency matrices. It provides flexibility for different methods:
\itemize{
\item The "vote" method considers an edge present if it appears in at least 75\% of the input matrices.
\item The "union" method includes any edge present in any matrix in the consensus matrix.
\item The "INet" method combines weighted matrices by normalizing the values and creating a consensus network
using the \code{consensusNet} function.
}
}
