% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_Atlas.R
\name{download_Atlas}
\alias{download_Atlas}
\title{Download and Load an RDS File from a URL}
\usage{
download_Atlas(file_url)
}
\arguments{
\item{file_url}{Character; URL of the RDS file to download.}
}
\value{
An R object loaded from the downloaded RDS file.
}
\description{
Downloads an RDS file from a specified URL and reads its contents into R.
We used it for https://www.singlecellatlas.org
}
\details{
This function uses \pkg{httr} to perform the download. The RDS file is read
directly from a raw connection without saving to disk. An internet connection
is required.

If the download fails (e.g., invalid URL, server error), an informative error
message is returned.
}
\examples{
url <- paste0(
    "https://www.dropbox.com/s/r8qwsng79rhp9gf/",
    "SCA_scRNASEQ_TISSUE_WHOLE_BLOOD.RDS?dl=1"
)
atlas_data <- download_Atlas(url)
}
