% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_Atlas.R
\name{download_Atlas}
\alias{download_Atlas}
\title{Download and Load an RDS File from a URL}
\usage{
download_Atlas(file_url)
}
\arguments{
\item{file_url}{A character string specifying the URL of the RDS file.}
}
\value{
The contents of the RDS file as an R object.
}
\description{
This function downloads an RDS file from a given URL and reads it into R.
}
\examples{
\dontrun{
data <- download_Atlas("https://example.com/data.rds")
}
}
