% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earlyj.R
\name{earlyj}
\alias{earlyj}
\title{Modify Row Names of Matrices and Combine Them}
\usage{
earlyj(matrix_list)
}
\arguments{
\item{matrix_list}{A list of matrices. Each matrix will have its row names modified.}
}
\value{
A combined matrix with modified row names.
}
\description{
This function modifies the row names of each matrix in a list by appending
the matrix index (e.g., "_m1", "_m2", ...) to the row names. It then combines
all the modified matrices into a single matrix using \code{rbind}.
}
\examples{
\dontrun{
combined_matrix <- earlyj(list(matrix1, matrix2, matrix3))
}
}
