% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adjacency.R
\name{generate_adjacency}
\alias{generate_adjacency}
\title{Generate Adjacency Matrices from a List of Data Frames}
\usage{
generate_adjacency(
  df_list,
  nCores = BiocParallel::bpworkers(BiocParallel::bpparam())
)
}
\arguments{
\item{df_list}{A list of data frames, each containing three columns:
\itemize{
\item First column: Gene1 (character)
\item Second column: Gene2 (character)
\item Third column: Weight (numeric)
}}
}
\value{
A list of adjacency matrices, one for each data frame in \code{df_list}.
}
\description{
This function generates adjacency matrices based on the information provided in
a list of data frames. Each data frame is expected to have three columns:
the first and second columns represent gene pairs, and the third column represents
the weight of the interaction between the genes.
}
\details{
The function creates an adjacency matrix for each data frame in \code{df_list} by placing
the interaction weights in the correct positions based on the gene pairings. The
diagonal of the resulting matrices is set to zero to ensure no self-interactions are included.
}
\examples{

# Example data frame
df1 <- data.frame(Gene1 = c("GeneA", "GeneB"), Gene2 = c("GeneB", "GeneC"), Weight = c(0.5, 0.8))
df_list <- list(df1)
adjacency_matrices <- generate_adjacency(df_list)

}
