% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_networks.R
\name{infer_networks}
\alias{infer_networks}
\title{Infer Gene Regulatory Networks Using Different Methods}
\usage{
infer_networks(
  count_matrices_list,
  method = "GENIE3",
  adjm = NULL,
  nCores = (BiocParallel::bpworkers(BiocParallel::bpparam()) - 1)
)
}
\arguments{
\item{count_matrices_list}{A list of matrices containing gene expression data (rows as genes and columns as samples).}

\item{method}{A character string specifying the method to use for inferring the network. Options are:
"GENIE3", "GRNBoost2", "ZILGM", "JRF", and "PCzinb". Default is "GENIE3".}

\item{adjm}{An optional adjacency matrix to be used in certain methods (e.g., for PCzinb). Default is \code{NULL}.}

\item{nCores}{An integer specifying the number of cores to use for parallel computation. Default is the number of cores minus one.}
}
\value{
A list containing the inferred networks. The exact structure depends on the method chosen.
}
\description{
This function infers gene regulatory networks (GRNs) from a list of count matrices using
several different methods such as GENIE3, GRNBoost2, ZILGM, JRF, or PCzinb.
}
\details{
\itemize{
\item "GENIE3" uses the GENIE3 algorithm to infer the GRN.
\item "GRNBoost2" uses the GRNBoost2 algorithm for network inference.
\item "ZILGM" uses the ZILGM method for network inference and also returns lambda values.
\item "JRF" uses the JRF method, normalizing the input matrices before applying the model.
\item "PCzinb" uses the PCzinb method for GRN inference.
}
}
\examples{
\dontrun{
result <- infer_networks(count_matrices_list = list(matrix1, matrix2), method = "GENIE3")
}
}
