% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCzinb_internal.R
\name{nb.loglik.dispersion}
\alias{nb.loglik.dispersion}
\title{Log-likelihood of  negative binomial model, for a fixed
dispersion parameter}
\usage{
nb.loglik.dispersion(zeta, Y, mu)
}
\arguments{
\item{zeta}{a vector, the log of the inverse dispersion parameters of the
negative binomial model}

\item{Y}{a vector of counts}

\item{mu}{a vector of mean parameters of the negative binomial}
}
\value{
the log-likelihood of the model.
}
\description{
Given a unique dispersion parameter and a set of counts, together with a
corresponding set of mean parameters,
this function computes the sum of the log-probabilities of the counts under
the NB model. The dispersion parameter is provided to the function through
zeta = log(theta), where theta is sometimes called the inverse dispersion
parameter.
}
