% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCzinb_internal.R
\name{nb.loglik.regression.gradient}
\alias{nb.loglik.regression.gradient}
\title{Gradient of the  log-likelihood of the NB regression model}
\usage{
nb.loglik.regression.gradient(
  alpha,
  Y,
  A.mu = matrix(nrow = length(Y), ncol = 0),
  C.theta = matrix(0, nrow = length(Y), ncol = 1)
)
}
\arguments{
\item{alpha}{the vectors of parameters a.mu concatenated}

\item{Y}{the vector of counts}

\item{A.mu}{matrix of the model (see Details, default=empty)}

\item{C.theta}{matrix of the model (see Details, default=zero)}
}
\value{
The gradient of the log-likelihood.
}
\description{
This function computes the gradient of the log-likelihood of a NB
regression model given a vector of counts.
}
\details{
The regression model is described in
\code{\code{nb.loglik.regression}}.
}
\seealso{
\code{\code{nb.loglik.regression}}
}
\keyword{internal}
