% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PCzinb_internal.R
\name{PCzinb}
\alias{PCzinb}
\alias{PCzinb,SummarizedExperiment-method}
\alias{PCzinb,matrix-method}
\alias{PCzinb,SingleCellExperiment-method}
\title{Generic functions for scGraphVerse}
\usage{
PCzinb(x, ...)

\S4method{PCzinb}{SummarizedExperiment}(x, whichAssay = "processed", ...)

\S4method{PCzinb}{matrix}(
  x,
  method = c("poi", "nb", "zinb0", "zinb1"),
  alpha = 2 * pnorm(nrow(x)^0.2, lower.tail = FALSE),
  maxcard = 2,
  extend = TRUE,
  nCores = 1
)

\S4method{PCzinb}{SingleCellExperiment}(x, whichAssay = "processed", ...)
}
\arguments{
\item{x}{the matrix of counts (n times p) or a SummarizedExperiment
containing such matrix (transposed).}

\item{...}{Arguments to pass to the matrix method.}

\item{whichAssay}{The assay to use as input for the matrix method.}

\item{method}{the algorithm used to estimate the graph: \code{poi}, \code{nb}, \code{zinb0},
or \code{zinb1}. See details below.}

\item{alpha}{the significant level of the tests}

\item{maxcard}{the uper bound of the cardinality of the conditional sets K}

\item{extend}{if TRUE it considers the union of the tests, otherwise it
considers the intersection.}
}
\value{
An adjacency matrix or SummarizedExperiment with adjacency matrix

if x is a matrix, the estimated adjacency matrix of the graph; if x
is a SummarizedExperiment, a SummarizedExperiment object with the adjacency
matrix as metadata.
}
\description{
This file contains all generic function definitions for the scGraphVerse package.
These must be defined before any setMethod calls.
}
\examples{
library(SummarizedExperiment)
se <- SummarizedExperiment(matrix(rpois(50, 5), ncol = 10))
PCzinb(se, method = "poi")
mat <- matrix(rpois(50, 5), nrow = 10)
PCzinb(mat, method = "poi")
library(SingleCellExperiment)
se <- SingleCellExperiment(matrix(rpois(50, 5), ncol = 10))
se <- PCzinb(se, method = "poi")
rowPair(se)
}
\keyword{Binomial}
\keyword{Binomial,}
\keyword{Negative}
\keyword{PC}
\keyword{Poisson,}
\keyword{Structure}
\keyword{This}
\keyword{Zero-Inflated}
\keyword{adapted}
\keyword{algorithms}
\keyword{and}
\keyword{assumptions}
\keyword{count}
\keyword{data}
\keyword{different}
\keyword{distributional}
\keyword{for}
\keyword{function}
\keyword{including}
\keyword{internal}
\keyword{learning}
\keyword{models.}
\keyword{performs}
\keyword{structure}
\keyword{using}
\keyword{various}
