% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC Curve for a List of Matrices}
\usage{
plotROC(matrices_list, ground_truth, plot_title, is_binary = FALSE)
}
\arguments{
\item{matrices_list}{A list of matrices. Each matrix is either binary or weighted.
For binary matrices, values are assumed to be either 0 or 1.}

\item{ground_truth}{A square matrix representing the ground truth. It contains binary values
where the upper triangle represents the interactions (1) or lack of interactions (0).}

\item{plot_title}{A string to be used as the title for the ROC curve plot.}

\item{is_binary}{A boolean indicating whether the matrices in the list are binary.
Default is FALSE, indicating that the matrices are weighted.}
}
\value{
A data frame containing the AUC values for each matrix in the list.
}
\description{
This function takes a list of matrices and their corresponding ground truth
to calculate and plot the Receiver Operating Characteristic (ROC) curve for
each matrix. The function supports both binary and weighted matrices. If the
matrices are binary, the function computes the True Positive Rate (TPR) and
False Positive Rate (FPR) for each matrix. If the matrices are weighted,
the function computes the ROC curve using the predicted values from the weighted matrices.
}
\details{
The function computes the ROC curve for each matrix, calculates the AUC value,
and plots the ROC curves. For binary matrices, the AUC is calculated based on the
True Positive Rate (TPR) and False Positive Rate (FPR). For weighted matrices,
the AUC is computed using the continuous predicted values. The function supports
multiple matrices in the input list and generates a dynamic plot for comparison.
}
\examples{
\dontrun{
# Example of plotting the ROC curve for a list of weighted matrices
plotROC(matrices_list = list(mat1, mat2), ground_truth = truth_matrix, 
        plot_title = "ROC Curve for Weighted Matrices", is_binary = FALSE)
}

}
