% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCzinb_internal.R
\name{pois.wald}
\alias{pois.wald}
\title{Structure learning with Poisson models}
\usage{
pois.wald(X, maxcard, alpha, extend, nCores = 1)
}
\arguments{
\item{X}{the matrix of counts (n times p).}

\item{maxcard}{the uper bound of the cardinality of the conditional sets K}

\item{alpha}{the significant level of the tests}

\item{extend}{if TRUE it considers the union of the tests, otherwise it
considers the intersection.}
}
\value{
the estimated adjacency matrix of the graph.
}
\description{
This function estimates the adjacency matrix of a Poisson model given a
matrix of counts, using the glm function.
}
\examples{
# Generate sample count data
set.seed(123)
n <- 50
p <- 4
X <- matrix(rpois(n * p, lambda = 3), nrow = n, ncol = p)

# Estimate network structure using Poisson model
adj_matrix <- pois.wald(X, maxcard = 1, alpha = 0.05, extend = TRUE)
}
