% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCzinb_internal.R
\name{Poisk2}
\alias{Poisk2}
\title{Structure learning with Poisson models using the Poisson K2 (PK2) algorithm}
\usage{
Poisk2(X, order, criterion = "BIC", maxcard)
}
\arguments{
\item{X}{the matrix of counts (n times p).}

\item{order}{the topological ordering of variables (names of nodes).}

\item{criterion}{the score function that measure the fitting of structures,
could be "AIC" or "BIC".}

\item{maxcard}{the uper bound of the cardinality of the parent sets.}
}
\value{
a list containing the estimated adjacency matrix of the graph and a
graphNEL object of the same graph.
}
\description{
This function finds the best fitting structure of a Poisson model given a
matrix of counts and topological ordering, using a given criterion ("AIC",
"BIC"). The PK2 algorithm is a modification of the K2 algorithm of Cooper and
Herskovits (1992) able to deal with Poisson data. See Nguyen et al. (2022)
for details.
}
\references{
Cooper, G. F. and Herskovits, E. (1992). A Bayesian method for the induction
of probabilistic networks from data. Machine learning, 9(4), 309–347.

Nguyen, Chiogna, Risso, Banzato (2022). Guided structure learning of
DAGs for count data. arXiv:2206.09754.
}
