% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscores.R
\name{pscores}
\alias{pscores}
\title{Performance Scores for Predicted Matrices}
\usage{
pscores(ground_truth, predicted_list, zero_diag = TRUE)
}
\arguments{
\item{ground_truth}{A square adjacency matrix representing the ground truth network.}

\item{predicted_list}{A list of square adjacency matrices to be evaluated against the ground truth.}

\item{zero_diag}{Logical; if TRUE, sets the diagonal of the ground truth matrix to zero.}
}
\value{
A list containing a data frame with computed statistics for each predicted matrix.
}
\description{
This function evaluates the performance of predicted adjacency matrices against a ground truth matrix.
It computes various classification metrics such as True Positive Rate (TPR), False Positive Rate (FPR),
Precision, F1-score, and Matthews Correlation Coefficient (MCC). Additionally, it visualizes the results
using a radar chart.
}
