% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selgene.R
\name{selgene}
\alias{selgene}
\title{Extract Top Highly Expressed Genes for a Specific Cell Type in a Seurat Object}
\usage{
selgene(seurat_object, cell_type, top_n = NULL)
}
\arguments{
\item{seurat_object}{A Seurat object containing RNA expression data.}

\item{cell_type}{A character string specifying the cell type of interest.}

\item{top_n}{An integer indicating the number of top expressed genes to select.}
}
\value{
A character vector of the selected top \code{n} highly expressed genes.
}
\description{
This function identifies and extracts the top \code{n} most highly expressed genes
for a given cell type from the RNA assay in a Seurat object. It computes the
average expression of each gene across all cells belonging to the specified
cell type and selects the highest-expressed ones.
}
\details{
The function first validates the presence of the specified \code{cell_type} in the
metadata and extracts the corresponding cells. It then retrieves the normalized
gene expression matrix from the \code{"RNA"} assay and calculates the mean expression
of each gene across the selected cells. The genes are sorted in descending order
based on their average expression, and the top \code{n} genes are returned.
}
\examples{
\dontrun{
selected_genes <- pathg(my_seurat_object, cell_type = "T_cells", top_n = 20)
print(selected_genes)
}

}
