% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrize.R
\name{symmetrize}
\alias{symmetrize}
\title{Symmetrize a List of Matrices}
\usage{
symmetrize(matrix_list, weight_function = "mean")
}
\arguments{
\item{matrix_list}{A list of matrices to be symmetrized. Each matrix should be square
and contain numeric values.}

\item{weight_function}{A string specifying the function to combine values at (i, j) and (j, i).
The default is "mean". Other options include "max", or any other function that can be
applied to a vector of two numeric values (e.g., "min").}
}
\value{
A list of symmetrized matrices. Each matrix will have identical values at positions
(i, j) and (j, i), computed using the specified weight function.
}
\description{
This function takes a list of matrices and symmetrizes each matrix by ensuring that
for each pair of off-diagonal elements (i, j) and (j, i), their values are the same.
The function allows you to specify how to combine the values at (i, j) and (j, i)
using a weight function, such as the "mean" or "max" of the two values.
}
\details{
For each matrix in the input list, the function processes every pair of off-diagonal elements
and applies the chosen \code{weight_function} to symmetrize the matrix. If one of the values at
(i, j) or (j, i) is zero, the function will use the non-zero value. If both are non-zero,
the function will apply the \code{weight_function} to the two values.
}
\examples{
\dontrun{
# Example list of matrices
mat1 <- matrix(c(0, 2, 3, 4), nrow = 2, ncol = 2)
mat2 <- matrix(c(0, 5, 6, 0), nrow = 2, ncol = 2)
matrix_list <- list(mat1, mat2)

# Symmetrize with "mean" function
symmetrized_matrices <- symmetrize(matrix_list, weight_function = "mean")
}

}
