% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCzinb_internal.R
\name{zinbOptimizeDispersion}
\alias{zinbOptimizeDispersion}
\title{(ZINB) model. The ZINB distribution is parametrized by three
parameters: the mean value and the dispersion of the negative binomial
distribution, and the probability of the zero component.}
\usage{
zinbOptimizeDispersion(mu, logitPi, Y, n)
}
\arguments{
\item{mu}{the vector mean  of the negative binomial}

\item{logitPi}{the vector of logit of the probabilities of the zero component}

\item{Y}{the vector of counts}

\item{n}{length of the returned vector}
}
\value{
A vector of length n with the optimized dispersion parameter values.
}
\description{
(ZINB) model. The ZINB distribution is parametrized by three
parameters: the mean value and the dispersion of the negative binomial
distribution, and the probability of the zero component.
}
\keyword{internal}
